merge feat/ui to dev

+++ src/main/java/com/trainticketbooking/app/Entities/Seat.java
@@ -13,7 +13,7 @@ public class Seat {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "seat_id")
-    private Long seatId;
+    private Integer seatId;
 
     // @ManyToOne
     // @JoinColumn(name = "compartment_id", nullable = true)
